fix infinite loop, don't use getStaticParams, don't use async page
idk what the deal is, maybe bc I'm not using fetch like next wants me
to?
But I kept getting infinite loops when navigating back. First it was the
waypoint page, but I removed the async keyword from the component bc I
wasnt using it. That worked, but then the system page had the bug
I removed the async keyword, and the async keyword from getSystemData,
as well as the getStaticParams. That did the trick, but idk what the
heck was going on.

I have static json, why is this so hard?